UHF-12681: Renamed folder
