+++ public/modules/custom/paatokset_ahjo_api/templates/block/block--meetings-calendar.html.twig
@@ -1,11 +1,10 @@
 {%
   set classes = [
     'meetings-calendar',
-    view_mode ? 'meetings-calendar--view-mode-' ~ view_mode|clean_class,
   ]
 %}
 
 {% embed "@hdbt/misc/lead-in.twig" with { content: meeting_calendar_description } %}{% endembed %}
 <div {{ attributes.addClass(classes) }}>
-  <div id="meeting-calendar"></div>
+  <div id="meetings-calendar"></div>
 </div>
