+++ public/themes/custom/hdbt_subtheme/src/js/react/apps/meetings-calendar/helpers/date.ts
@@ -1,3 +1,4 @@
+/** Zero-pads a single-digit number to 2 digits, e.g. 7 -> '07'. */
 const pad = (value: number): string => String(value).padStart(2, '0');
 
 /**
@@ -9,17 +10,16 @@ export const parseLocalDate = (dateString: string): Date => {
   return new Date(year, month - 1, day);
 };
 
-/** Short day label without leading zeroes, e.g. '29.7' (matches previous dayjs 'D.M' format). */
+/** Short day label without leading zeroes, e.g. 29.7 */
 export const formatDayShort = (date: Date): string => `${date.getDate()}.${date.getMonth() + 1}`;
 
-/** Zero-padded full date, e.g. '29.07.2026' (matches previous dayjs 'DD.MM.YYYY' format). */
+/** Full date, e.g. '29.07.2026' */
 export const formatDayFull = (date: Date): string =>
   `${pad(date.getDate())}.${pad(date.getMonth() + 1)}.${date.getFullYear()}`;
 
 /**
  * Subtracts whole months from a date, clamping to the target month's last day
- * (e.g. 31 July - 3 months = 30 April). Plain `Date#setMonth` would instead roll
- * over into the following month whenever the target month is shorter.
+ * (e.g. 31 July - 3 months = 30 April).
  */
 export const subtractMonthsClamped = (date: Date, months: number): Date => {
   const targetMonthIndex = date.getMonth() - months;
@@ -30,6 +30,6 @@ export const subtractMonthsClamped = (date: Date, months: number): Date => {
   return new Date(year, month, day);
 };
 
-/** Steps a calendar month forward/backward, always landing on the 1st (never overflows). */
+/** Steps a calendar month forward/backward, always landing on the 1st. */
 export const addCalendarMonths = (firstOfMonth: Date, delta: number): Date =>
   new Date(firstOfMonth.getFullYear(), firstOfMonth.getMonth() + delta, 1);
