+++ public/themes/custom/hdbt_subtheme/theme-builder.mjs
@@ -12,7 +12,7 @@ const outDir = path.resolve(__dirname, 'dist');
 const reactApps = {
   'allu-decisions-search': './src/js/react/apps/allu-decisions-search/index.tsx',
   'decisions-search': './src/js/react/apps/decisions-search/index.tsx',
-  'meeting-calendar': './src/js/react/apps/meeting-calendar/index.tsx',
+  'meetings-calendar': './src/js/react/apps/meetings-calendar/index.tsx',
 };
 
 // Vanilla JS files.
